new BookKeeperException and fix SyncCallbackUtils
